+++ NMSReportingSuite/src/main/java/com/beehyv/nmsreporting/entity/KilkariThematicContentReportDto.java
@@ -1,4 +1,5 @@
 package com.beehyv.nmsreporting.entity;
+import java.math.BigDecimal;
 
 
 /**
@@ -13,7 +14,7 @@ public class KilkariThematicContentReportDto {
     private String messageWeekNumber;
     private Long uniqueBeneficiariesCalled;
     private Long callsAnswered;
-    private Long minutesConsumed;
+    private BigDecimal minutesConsumed;
 
 
     public Integer getId() {
@@ -56,11 +57,11 @@ public class KilkariThematicContentReportDto {
         this.callsAnswered = callsAnswered;
     }
 
-    public Long getMinutesConsumed() {
+    public BigDecimal getMinutesConsumed() {
         return minutesConsumed;
     }
 
-    public void setMinutesConsumed(Long minutesConsumed) {
+    public void setMinutesConsumed(BigDecimal minutesConsumed) {
         this.minutesConsumed = minutesConsumed;
     }
 }
